requirements.txt
# Requirements for MedGemma Fine-tuning on CTRate
# Install with: pip install -r requirements.txt

# Core dependencies
torch>=2.0.0
transformers>=4.40.0
peft>=0.10.0
bitsandbytes>=0.43.0
trl>=0.9.0
accelerate>=0.25.0

# Dataset and data processing
datasets>=2.18.0
pillow>=10.0.0

# Medical imaging
pydicom>=2.4.0
nibabel>=5.2.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0

# Training utilities
tensorboard>=2.15.0
wandb>=0.16.0

# HuggingFace Hub
huggingface-hub>=0.21.0
